app/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
babel==2.16.0
billiard==4.2.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
Django==5.1.1
django-celery-beat==2.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-crontab==0.7.1
django-css-inline==0.0.6
django-dbbackup==4.2.1
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-mailbox==4.10.1
django-rest-swagger==2.2.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docxcompose==1.4.0
docxtpl==0.18.0
drf-yasg==1.21.7
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.8
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
kombu==5.4.0
lxml==5.3.0
MarkupSafe==2.1.5
mysqlclient==2.2.4
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.9.0
pyparsing==3.1.4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
XlsxWriter==3.2.0
